Added Messages Features
