+++ src/index.html
@@ -29,7 +29,9 @@
                     <div class="row" id="messages">
                         This is where the messages will go.
                     </div>
-                    <div class="row" id="message-list"></div>
+                    <div class="row" id="message-list">
+                        This is where the messages list will go.
+                    </div>
                     
 
                     <div class="row" id="tasks">
