+++ src/scripts/Nutshell.js
@@ -1,7 +1,13 @@
 import {LogOutButton} from "./auth/LogoutButton.js"
+import { messages} from './chat/messageList.js'
+import { messageForm } from './chat/messageInput.js'
+
+
 
 export const Nutshell = () => {
 
     LogOutButton()
-      // Render all your UI components here
+    messages()
+    messageForm()
+          // Render all your UI components here
 }
