+++ src/styles/main.css
@@ -1,3 +1,4 @@
 * {
     border: 1px solid black;
 }
+
